Sync Translate Language for Dashboard and Profile
