+++ src/components/Menu2/index.jsx
@@ -6,12 +6,9 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import i18n from 'translations/i18n';
 import { withTranslation } from 'react-i18next';
 import { faArrowLeft } from '@fortawesome/free-solid-svg-icons/faArrowLeft';
 import { faUser } from '@fortawesome/free-solid-svg-icons/faUser';
-import { Dropdown } from 'react-bootstrap';
-import { faGlobe } from '@fortawesome/free-solid-svg-icons/faGlobe';
 
 class Menu2 extends React.Component {
   constructor(props) {
@@ -48,9 +45,7 @@ class Menu2 extends React.Component {
   render() {
     let { dataMenu } = this.state;
     const { t } = this.props;
-    const listLanguages = Object.keys(i18n.options.resources).map(function (key) {
-      return { language: key, title: i18n.options.resources[key].title };
-    });
+
     return (
       <nav>
         <div className="py-1 px-3 item_menu item_menu_home">
@@ -92,31 +87,6 @@ class Menu2 extends React.Component {
             </div>
           );
         })}
-        <div className="position-absolute bottom-0 mb-3 border-top w-100 py-1 button-language">
-          <Dropdown className="pt-2 ">
-            <Dropdown.Toggle variant="dark" id="dropdown-basic" className="bg-transparent border-0">
-              <FontAwesomeIcon icon={faGlobe} /> {this.language ?? 'English'}
-            </Dropdown.Toggle>
-
-            <Dropdown.Menu>
-              {listLanguages.map((item, index) => {
-                return (
-                  <Dropdown.Item
-                    key={index}
-                    href="#"
-                    className=""
-                    onClick={() => {
-                      i18n.changeLanguage(item.language);
-                      this.setState((this.language = item.title));
-                    }}
-                  >
-                    {item.title}
-                  </Dropdown.Item>
-                );
-              })}
-            </Dropdown.Menu>
-          </Dropdown>
-        </div>
       </nav>
     );
   }
