+++ src/translations/dk/common.json
@@ -57,7 +57,7 @@
   "txt_title_calendar": "Planlæg",
   "txt_title_projects": "Projekter",
   "txt_title_campaigns": "List kampagner",
-  "txt_title_content": "Liste indlæg",
+  "txt_title_content": "Listeprojekt",
   "txt_title_channels": "Tilslut en kanal",
   "txt_title_digital_assets": "Digitale aktiver",
   "txt_title_new_calendar": "Ny tidsplan",
