+++ src/translations/en/common.json
@@ -57,7 +57,7 @@
   "txt_title_calendar": "Schedule",
   "txt_title_projects": "Projects",
   "txt_title_campaigns": "List Campaigns",
-  "txt_title_content": "List posts",
+  "txt_title_content": "List Project",
   "txt_title_channels": "Connect a Channel",
   "txt_title_digital_assets": "Digital Assets",
   "txt_title_new_calendar": "New Schedule",
