+++ src/translations/es/common.json
@@ -57,7 +57,7 @@
   "txt_title_calendar": "Calendario",
   "txt_title_projects": "Proyectos",
   "txt_title_campaigns": "Listar campañas",
-  "txt_title_content": "Listar publicaciones",
+  "txt_title_content": "Listar Proyectos",
   "txt_title_channels": "Conectar un canal",
   "txt_title_digital_assets": "Recursos digitales",
   "txt_title_new_calendar": "Nuevo horario",
