+++ src/translations/ua/common.json
@@ -57,7 +57,7 @@
   "txt_title_calendar": "Розклад",
   "txt_title_projects": "Проекти",
   "txt_title_campaigns": "Список кампаній",
-  "txt_title_content": "Список дописів",
+  "txt_title_content": "Список проекту",
   "txt_title_channels": "Підключити канал",
   "txt_title_digital_assets": "Цифрові активи",
   "txt_title_new_calendar": "Новий розклад",
