+++ src/translations/vi/common.json
@@ -57,7 +57,7 @@
   "txt_title_calendar": "Lịch",
   "txt_title_projects": "Các Dự Án",
   "txt_title_campaigns": "Danh sách các Chiến Dịch",
-  "txt_title_content": "Danh sách các Bài Đăng",
+  "txt_title_content": "Danh sách các Dự Án",
   "txt_title_channels": "Liên kết các Kênh",
   "txt_title_digital_assets": "Tài Nguyên Kỹ Thuật Số",
   "txt_title_new_calendar": "Lịch Mới",
